add casts for gcc
git-svn-id: https://svn.r-project.org/R/trunk@72154 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/main/array.c
@@ -489,7 +489,7 @@ static SEXP do_lengths_long(SEXP x, SEXP call, SEXP rho)
     x_len = dispatch_xlength(x, call, rho);
     PROTECT(ans = allocVector(REALSXP, x_len));
     for (i = 0, ans_elt = REAL(ans); i < x_len; i++, ans_elt++)
-        *ans_elt = getElementLength(x, i, call, rho);
+        *ans_elt = (double) getElementLength(x, i, call, rho);
     UNPROTECT(1);
     return ans;
 }
